Meta bodies are now always instantiated one at a time by keeping seperate queues for meta bodies waiting to execute and actual statements of the current meta body.  This should fix several errors with decomposition of ports as well as allow for future simplification of the code for executing connect statements.  However, there is still a lot of work that needs to be done with regard to the usability of the debug prompt during instantiation.
Currently, instantiation time type checking of ports as they are connected
has removed a lot of the need for run time range checking during
communications.  This type checking has been added to ports that are
decomposed through the value_union structure, verifying that the return
type of the decomposition function matches exactly the decomposed type.
Currently this check is executed every time a value_union is created,
which is probably more times than the check actually needs to be run.