chpconv now can recursively output all subprocesses of a single process, and can either place all output into a single file, or each process into its own file.
chpconv also still had a lot of bugs that should now be fixed.  One outstanding
bug that does not have an easy fix is that chpconv cannot distinguish between
wires that are bridged internally to a process versus wires that are bridged
externally to a process.  However, wires that are connected to a rail are
simply marked as connected to a rail rather than actually being bridged, so
this bug generally will not pop up.