Fix to allow access to more decomposed/sliced ports from the debug prompt.
The syntax printed on the debug prompt when setting the view to a process
using a default wired port is now mostly correct.

Removal of a large amount of unused code from the routines for printing to
the debug prompt.

Fix for a bug where a changing wire value could trigger at most one
counter rule to fire.